+++ src/app/services/horostorage/horostorage.service.ts
@@ -3,6 +3,7 @@ import { ProcessName } from 'src/app/process/enum/process';
 import { DirectionMethod } from 'src/app/process/enum/direction-method';
 import { ArcToDateMethod } from 'src/app/process/enum/arc-to-date-method';
 import { ProfectionArcToDateMethod } from 'src/app/process/enum/profection-arc-to-date-method';
+import { DailyDirectionMethod } from 'src/app/process/enum/daily-direction-method';
 import { DeepReadonly } from 'src/app/type/interface/deep-readonly';
 import {
   HoroRequest,
@@ -140,6 +141,7 @@ export class HoroStorageService {
         direction_method: DirectionMethod.SemiArc,
         arc_to_date_method: ArcToDateMethod.DegreePerYear,
         profection_arc_to_date_method: ProfectionArcToDateMethod.TrueSolarArc,
+        daily_direction_method: DailyDirectionMethod.SemiArcZodiacal,
       });
     }
   }